Fix JS linter issues.
