+++ assets/js/books.js
@@ -1,0 +1,9 @@
+class Book {
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+    this.id = 0;
+  }
+}
+
+export default Book;
